requirements.txt
requests==2.31.0
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.2
pydantic_core==2.14.5
pydantic-settings==2.1.0
aws-lambda-powertools==2.28.0
jmespath==1.0.1
pytest==7.4.3
python-dotenv==1.0.0
uvicorn==0.24.0
loguru==0.7.2
pytest==7.4.3